+++ packages/ui/CHANGELOG.md
@@ -39,6 +39,7 @@ No significant changes.
 
 - `ui/FloatingLayerDecorator` to render floating node properly
 - `ui/Touchable' to handle touch related events only for valid targets
+- `ui/ViewManager` to set `state.entering` to `false` when `noAnimation` prop is changed to `true`
 
 ## [4.0.0-alpha.1] - 2021-02-24